[compoundRepair] total 초기화해줌 usualAmount, goalAmount  무조건 하루당으로 계산. 그러면 Habit에서 자료형이 int가 아니라 double이 되어야함 그래서 딸려있는 코드들도 int에서 double로 바꿔줬는데 알 수 없는 오류발생

+++ rich_habit/lib/habit.dart
@@ -9,10 +9,10 @@ class Habit {
   bool isTrigger;
 
   bool usualIsWeek; // 평상시 횟수 단위 Daily = false, Weekly = true
-  int usualAmount; //평상시 습관 양
+  double usualAmount; //평상시 습관 양
 
   bool goalIsWeek; // 목표 횟수 단위 Daily = false, Weekly = true
-  int goalAmount; // 목표 습관 양
+  double goalAmount; // 목표 습관 양
 
   int nowAmount = 0; // 현재 습관 양
 
@@ -76,7 +76,7 @@ class Habit {
     this.goalIsWeek = isWeek;
   }
 
-  void modifyGoalAmount(int goalAmount) {
+  void modifyGoalAmount(double goalAmount) {
     this.goalAmount = goalAmount;
   }
